[PATCH] Workarounds for glXGetProcAddress quirks
Some people using the Nvidia binary driver on their systems are having
trouble with linking against libGL since glXGetProcAddress is not
defined. These systems seem to often have the Mesa GL headers
installed, so I can't detect this problem at compile time. Apparently
glXGetProcAddressARB is part of the Linux OpenGL ABI, so it should be
safe to always use that.

I've implemented this as a workaround in the Makefile in the hope that
the situation will resolve itself in the various libGL implementations
eventually and I'll be able to just use the headers reliably. If I
really have to make it a portable binary, I'll just have to do some
dlopen magic in the future.

Signed-off-by: Tyrann <tyrann@disenchant.net>