[PATCH] Hunk_TempAllocExtend
This patch implements a new memory allocation feature, which allows us
to use the "Temp" hunk as an expandable blob of storage space. The
first user of this will be some "string tree" functions, which build
up an RB tree of strings which will be used for sorting file lists.

We could just use Z_Malloc, but the zone isn't well suited to this
type of allocation (potentially hundreds of small struct + string
allocations) and is probably too small.

Signed-off-by: Tyrann <tyrann@disenchant.net>