[PATCH] String Trees
This defines structures for an RB tree of strings. The intended use of
this is for quick insertion sort of filenames for tab completion and
any other uses. We use the temp hunk for (reasonably) efficient
allocations - at least a little better than loading up the zone with
hundreds of small allocations.

Signed-off-by: Tyrann <tyrann@disenchant.net>