[PATCH] Add ST_MaxMatch function
Introduce a new string tree function, ST_MaxMatch. This will be our main
helper function for tab completion.

Given a partial string, this function will search a string tree for strings
who's prefix matches the partial string. It will then return a string which is
the maximum common prefix of the matching strings (i.e. can be longer than the
given partial string.

For now, if the match is unique, the returned string has a space appended. I'm
not sure if I want to keep it like that as it feels like a bit of a hack. It
will do for now.

Signed-off-by: Tyrann <tyrann@disenchant.net>