[PATCH] Command argument completion callback
Add an argument completion callback to the command struct. Defaults to NULL.

If set, when tab completion is requested on the argument of a valid command,
the callback will be invoked, passing in the partial argument. The callback
then returns a (possibly partially) completed string.

The initial targets for this will be commands like map, playdemo, etc.

Signed-off-by: Tyrann <tyrann@disenchant.net>