[PATCH] Introduce container_of macro
Introduce a new macro "container_of" to convert a member point to a point to
the containing struct with relative type saftey. Stolen from Linux. I'd rather
this was in a different header to "qtypes.h", but that's the best place for it
at the moment, without creating a new header.

Replace the cvar_for_completion macro and add helper functions stree_entry and
completion_entry to remove the rb_entry macro. This ends up looking a bit
neater.

Signed-off-by: Tyrann <tyrann@disenchant.net>