[PATCH] Cmd STree
Replace the linked list for the cmd_function struct with an STree. All uses of
this list were to find a specific function, so they've all been replaced by
Cmd_FindCommand. This means I was able to rip out the ->next member straight
away.

Also, constified stree_node->string in the process.

Signed-off-by: Tyrann <tyrann@disenchant.net>