[PATCH] STree ForEach macro
Implement a ForEach style macro for iterating over string trees. Comments on
the usage and implementation in the code. I hadn't seen a construct like this
before - it was an interesting challenge...

The stack required for keeping state information is allocated dynamically, so
a special "break" command is also provided to make sure the memory is freed
properly. The _After version allows you to start iterating from a given entry
in the tree. This was implemented specially for one case where it was required
in NQ for cvars. Not sure if it will get a lot of use elsewhere...

Signed-off-by: Tyrann <tyrann@disenchant.net>