[PATCH] world.c cleanups
Mostly whitespace cleanups in the world.c's, and a little one in pmovetst.c
since it's largely copied from world.c. This is mainly to keep the differences
between NQ/QW versions as small as possible.

The only (minor) code changes are a little cleanup of SV_TestEntityPosition
(should be semantically the same) and changing the "backup past 0" message to
be developer only.

Signed-off-by: Tyrann <tyrann@disenchant.net>