[PATCH] Allow up to 65520 clipnodes
The BSP file format's original design limits us to having 32767 clipnodes by
having the offsets stored as 16-bit signed values. Negative values are to
signify leaf nodes and refer to the CONTENTS_xxx flags defined in bspfile.h.

Since only a handful of values are required for the CONTENTS_xxx flags, we can
make use of all the space on the negative side be being a bit more careful
about how we handle these offsets.

Essentially, we can treat the offset as unsigned and instead of checking for a
value < 0 to signify contents, we check that the value large enough to be
contents. This is binary compatible with existing maps.

I added a helper function clipnode_child() to extract a 32-bit signed integer
value from the dclipnode_t structure. If the child index is large enough to be
a contents flag, we return the negative value. Otherwise, we return the index
between 0 and 65520 (0xfff0).

Hopefully my attempt at updating the asm code looks okay - this part of the
patch is the result of a 1-day crash course in x86 asm.

Signed-off-by: Tyrann <tyrann@disenchant.net>