[PATCH] Build system update
This is a fairly big change to the makefile system. It does two main things:

- It removes the need to edit the Makefile to build the different
executables. At the moment, it's still necessary to edit the Makefile to
change to/from a debug build or to exclude x86 assembly.

- It removes the need to make clean every time you want to build another
executable. However, the header dependencies are still absent, so you need
to be careful when editing headers, as the object files won't get refreshed
automatically (yet).

All object files now get placed in a 'build' directory, with each
application's object files kept in it's own subdirectory. We also no longer do
a recursive make invocation in each of the application directories.

Signed-off-by: Tyrann <tyrann@disenchant.net>