[PATCH] Add automatic dependency generation
This is a first pass at getting automatic file dependencies generated.

We use gcc's -M options to have it generate the makefile rules for each object
file. The rule for and object file OBJ is stored in the file .OBJ.d. The
dependecy file is automatically updated whenever the object file gets updated.

I think this can be cleaned up quite a bit (there seems to be a lot of
redundancy in there), but it's working right now and I want to get something
committed.

Signed-off-by: Tyrann <tyrann@disenchant.net>