[PATCH] Fix sprite cache error
Only alias models should be setting anything in the mod->cache.data field, and
only through Mod_Extradata(). Stop the sprite loading code from setting this
field. This fixes a bug that could occur if an alias model was to replace an
unloaded sprite in the model cache.

Also, set mod->flags back to zero in a couple of places for QW, same as NQ.

Signed-off-by: Tyrann <tyrann@disenchant.net>