[PATCH] CD audio refactoring
Begin some re-factoring of the CD audio code. There's a lot of commonality in
there, and I'd like to abstract the OS specific parts of that a little better
than it is now. This patch starts things off by moving the CD_f function into
a new cd_common.c file which is shared between Linux and MS Windows.

Some cleanups of the mciSendCommand calls in cd_win.c got pulled in as well. I
prefer the assignment above the if statement.

Signed-off-by: Tyrann <tyrann@disenchant.net>