[PATCH] Add function to check if CD is playing
Add a driver function for the generic code to call to see if the CD is still
playing. The windows driver uses a seperate message handler as opposed to the
Linux driver, which polls the drive for the status change. The Windows driver
keeps track of the state internally to implement the polling interface.

Now that both drivers behave the same, the CDAudio_Update function can be
pulled into the generic layer.

Signed-off-by: Tyrann <tyrann@disenchant.net>