[PATCH] Fix the CD SetVolume function
Somehow I missed the fact that we could have been setting the CD volume to
whatever random value happened to be on the stack (well, only if it was
between 0.0 and 1.0). Anyway, let's make it a bit clearer what's going on:

o We clamp the volume variable to values between 0.0 and 1.0
o Note that the driver SetVolume function may return a different value to the
one we passed in if it can't set that exact volume.
o In any case where we altered the value of bgmvolume, we need to call
Cvar_SetValue again so that the string is updated to match the float value.

Signed-off-by: Tyrann <tyrann@disenchant.net>