[PATCH] Add home directory gamedir overlay
If the HOME environment variable is set, include $HOME/.tyrquake/<gamedir> in
searchpath. This directory will also be used as the location for file writes
(config.cfg, glquake cachedir) as well if available and writeable.

Thanks to Alejandro Pulver for providing the original patch, which I used as a
basis for this implementation.

Signed-off-by: Tyrann <tyrann@disenchant.net>