[PATCH] Calculate BSP model checksum only once
When calculating the checksum of the various lumps in the BSP model, the
checksum is being calculated twice for the same data. Just cache the result
from the first pass and use that.

This patch also brings the qwsv code into line with the others and removes an
unused global variable "model_checksum".

Signed-off-by: Tyrann <tyrann@disenchant.net>