[PATCH] Fix hang in SCR_ModalMessage for X11
SCR_ModalMessage relies on Sys_SendKeyEvents to return a keypress in response
to the message. The however, X11 software renderer wasn't doing this. This
patch fixes the software X11 renderer (heh, probably should be in the input
driver), bringing it closer into line with the GLX renderer.

Bug reported by Steven A <stevenaaus@yahoo.com>.

Signed-off-by: Tyrann <tyrann@disenchant.net>