[PATCH] Fix connection state checking in NET_SendToAll()
Fix NET_SendToAll() so that it really skips non-connected clients. Rename the
cryptic state1 and state2 variables to something more indicative of their
purpose. While we're at it, change the type of the blocktime argument to
double (the comparison is against a double.)

Based on patch by O.Sezer <sezero@users.sourceforge.net>

Signed-off-by: Tyrann <tyrann@disenchant.net>