[PATCH] Implement stricter prog strings checking
When loading progs we calculate the memory used by the strings. Then, in
PR_SetString, we check for strings outside this range (not just on the
negative side). This causes edict strings to end up in the pr_strtbl too, so
need to make it a bit larger (this will be made to grow dynamically later on).

Signed-off-by: Tyrann <tyrann@disenchant.net>