[PATCH] Use FLT_MAX macros when bounds checking
Just in case some wierd model or other file is thrown at us where the surface
extents or similar are outside of (-9999, 9999), use the FLT_MAX macro to
ensure we catch everything. I'm yet to see this actually happen in a compiled
BSP, but I do remember being bitten in the qbsp compiler code a while back.

Signed-off-by: Tyrann <tyrann@disenchant.net>