[PATCH] Add two IP address options to NQ
Add the "-bindip" option. Enables the user to bind to a specific IP address on
a multi-homed host (rather than INADDR_ANY - all interfaces). Note that using
this option will prevent us from receiving broadcast packets, so server
discovery on the LAN will not work if the server uses this option.

Add the "-ip" option. Enables the user to specify an IP to use in certain
Quake network protocol packets. This is really just a work around for the
problems inherent when embedding the client/server IP addresses inside the
network protocol.

The specific example which drove this implementation is that of a multi-homed
server who needs to listen on INADDR_ANY in order to receive broadcast server
discovery requests but needs to know which of it's IP addresses it should
return to the client in the serverinfo reply. Really, the client should just
look at the source address of the serverinfo reply, but we need to stay
somewhat backwards compatible with old hosts.

Signed-off-by: Tyrann <tyrann@disenchant.net>