[PATCH] Tidy up IP address command line options
renamed the misleading -ip option in NQ (tyrquake-0.57) to -localip.
made the ip binding option to check for -ip and removed -bindip: now
it is similar to its quakeworld counterpart. moved the relevant
argument cheking to init procedures and added checks for invalid input,
bound the valid options to static variables. documented the options.

Signed-off-by: O.Sezer <sezero@users.sourceforge.net>
Signed-off-by: Tyrann <tyrann@disenchant.net>