[PATCH] Fix the 0 byte udp packet DoS bug
NQ checks to see if there were any bytes to be read and if there were 0 bytes,
just carying on without reading anything. It seems that the 0 byte packet was
blocking the queue and so a 0 byte recvfrom is needed to make sure it gets
removed.

(Fix comes from the quakeforge tree)

Signed-off-by: O.Sezer <sezero@users.sourceforge.net>
Signed-off-by: Tyrann <tyrann@disenchant.net>