[PATCH] Simplify direct sound buffer locking calls
The direct sound buffer lock function has optional arguments to return a
second locked region in the sound buffer for the case where the buffer
wraps. This was never used in our case, so just remove these arguments.

We never should have let the DWORD typedef leak outside the windows only
portions of the code, so clean that up while we're here - pbuf is type
agnostic.

Signed-off-by: Tyrann <tyrann@disenchant.net>