[PATCH] Fix pr_strings_size calculation
Well, it turns out that progs->numstrings is actually the number of bytes
dedicated to storage of the progs strings, not the number of strings.
Consequently, our attempt to calculate the strings size was completely bogus.
To avoid futher confusion, we'll rename the numstrings member in the progs
header.

Thanks to O.Sezer <sezero@users.sourceforge.net> for pointing this out :)

Signed-off-by: Tyrann <tyrann@disenchant.net>