[PATCH] Release mouse in NQ/vid_win.c when game loses key focus
Copy over the mouse grab/release code from QW's vid_win.c, which releases the
mouse cursor when the console is down or when the menu is active. This
interacts badly with the way VID_HandlePause() is done, so disable that for
now (I do plan to fix it again later though).

Signed-off-by: Tyrann <tyrann@disenchant.net>