[PATCH] Abstract DMA sound buffer locking
Add SNDDMA_LockBuffer() and SNDDMA_UnlockBuffer() functions to the generic
sound interface, to allow drivers to implement appropriate locking on their
DMA buffers if required. Move the #ifdef'd Windows specific code to do this
from the shared snd_mix.c and snd_dma.c files to the windows sound driver.

Thanks to O.Sezer <sezero@users.sourceforge.net> for the idea and for
providing the orginal patch which was the basis for this one.

Signed-off-by: Tyrann <tyrann@disenchant.net>