[PATCH] Reduce the severity of a map loading bug
This bug has been around forever, but is non-trivial to fix. Basically we need
proper error handling to be done when the map load is requested. Trickier than
it sounds because the map load is done asynchronously via the command
buffer. For now, just reduce the "hang" from 60s down to 15s.

Signed-off-by: Tyrann <tyrann@disenchant.net>