[PATCH] Query size of X11 gamma ramp
It seems since Xorg 7.4, the size of the gamma ramp is no longer 256
in all cases. So we will need to query the gamma size and use the
appropriate size when calling into the X11 API. Quake assumes a 256
valued gamma ramp througout the code, so we need to scale the ramp
passed down via VID_SetGammaRamp accordingly.