trivial: remove various redundant variable assignments
clang found a number of places where we redundantly set a variable to
some value and never read it back (before it is assigned
again). Remove those cases. A couple were cases of trying to silence
the compiler, but modern versions of gcc seem to have made some of
these unneccessary.