Remove some unneeded variables
clang pointed out a number of "unused" variables. These are "used" in
the sense that we do assign things to them, but the results are never
looked at.

As far as I can tell, all the cases included in this commit don't
point to any subtle bugs and should be safe to remove.