Save stack context in Host_Init() before flushing the command buffer.
In case one of the command's execution fails and causes a longjmp()
call such as by way of a Host_Error(), we would segfault because we
haven't saved the stack context/environment. This was the case with a
command line like "tyr-quake +connect localhost" without any servers
running. So, added a setjmp() call before Cbuf_Execute().

While we're at it, set host_initialized to true before exec of
quake.rc and we'll rename host_abortserver in NQ to host_abort, in
line with QW.

Modified version of patch by O.Sezer <sezero@users.sourceforge.net>

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>