protocol: wrap parsing of sound indexes
The index of sound numbers is by byte in the standard NQ protocol, but
can be done by short in alternative protocols. Wrap the reading of
sound numbers on the client side to keep the protocol specific changes
obvious and contained.

Note that "static" vs. "dynamic" sounds are treated separately as they
can have different types in particular protocols (e.g. BJP3).

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>