qw: Fix simulation rate limiting
Remove this seemingly harmless check for something silly like time going
backwards - it happens a lot. Very confusing because it doesn't happen on
debug builds, but only when code optimisation is enabled.

It's very possible for a time arguement of zero to be passed in here.
However, when opimisation is enabled the FP math gets a bit looser and
ordering of operations can shift such that the float passed in is a very
small negative number.

When that happens and we reset the oldrealtime, we get tricked into
running too many simulation frames and the game speed gets screwed up.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>