nq: handle visedicts similarly to qw
Make handling of cl_visedicts as similar as possible to qw. I am not yet
100% conviced that this has not introduced subtle bugs due to now taking a
copy of the entities in this list, but merging the behaviour is probably
the right way to go (and fix the bugs later).

Quickly played through a couple of random maps and nothing jumped out as
being obviously incorrect.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>