render: Pass the world entity down from R_Mirror
Blending the mirror texture over the scene - mirror surfaces are on
the world only, so the world entity is the correct one to pass down.

Maybe an obscure bug fix as I'm fairly sure that currententity would
otherwise have been off somewhere down the end of cl_visedicts from a
previous loop through the entity list.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>