draw: fix nonsense sentinel value on edge list
This sentinel is probably not needed, as per the comment, but I'm not familiar
enough yet to confirm. However, the value given (2000 << 24) would evaluate to
-805306368 due to the integer overflow.

The comment would indicate that the value is supposed to be large and positive
as the sorting is done in ascending ->u order. In any case the code still works
like this, but at least the sentinal value makes more sense.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>