model: remove alias group structure in sw alias models
Simplify model animations by removing the alias group concept and making the
structures more like the gl models. We lose the ability to have per-pose
intervals, but those never worked in glquake and it appears all the original
assets just use a fixed interval anyway.

We now have the same MAXALIASFRAMES limit as glquake, but I'm sure that can be
worked around later (for both renderers).

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>