zone: Add Cache_AllocPadded allocator
I want to be able to have objects in the cache which are are members of a
containing struct, so enable the cache allocation to specify that there is a
certain amount of padding below the memory address of the managed pointer.

Following patches for alias models will show the intended use case.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>