model: rename modelorg global and move to gl_rsurf.c
Rename modelorg to bmodelorg and move it to gl_rsurf.c where it is used
for the R_RecursiveWorldNode function. Make it static and add a comment
about it's intended use.

Now it's obvious that the modelorg global in gl_rmain was acutally
set but never used for GL alias model rendering.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>