common: constify common.[ch] function prototypes
Make as many arguments to functions in common.h as possible const so
the compiler knows we shouldn't be modifying some buffers/strings
passed down here.

Also fixed (but not tested) the COM_BlockSequenceCheckByte function in
QW/common/common.c which has been #if 0'd for a while. Probably won't
keep it around, but fixed for completeness.

I skipped COM_Parse in this patch, will be constified in the following
patch.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>