input: fix hang on "Are you sure you want to start a new game?" with sdl
When using SDL video/input, the game hangs when we select "New Game" from
the menu and the "Are you sure you want to start a new game?" prompt is
displayed, because we stop processing input events.

This is a bit of a hackish workaround, but will try to untangle this a bit
more later.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>