build: use the Apple OpenGL framework for non-X11 OS X targets
Although my SGL stub was building before, it was linked against the
X11 version of the GL libs, whereas libSDL was linked against the
apple version => random chaos!

Need to include <OpenGL/gl.h> instead of <GL/gl.h>, so use a
conditional define to do the (-DAPPLE_OPENGL in the Makefile). When
linking, use "-framework OpenGL" and not -lGL.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>