common: replace some use of SZ_Print with MSG_WriteString
SZ_Print is really meant for occasions where we need to build up a
string via strcat into the sizebuf - MSG_WriteString is more
appropriate in the cases fixed here, usually following a
MSG_WriteByte() anyway, so nothing to append to. Use the
MSG_WriteStringf variant where appropriate to reduce reliance on the
va() buffers.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>