render: fix r_fullbright in glquake
Been meaning to look into this for a long time - fix switching
r_fullbright on and off for glquake. Not the most efficient
implementation because we still calculate the dynamic light stuff in
the background, but not really important to optimise it too much I
guess. If I sort out the caching and invalidation we could optimise
more later.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>