model: add foreach_leafbit macro to iterate over one bits
Probably the bigger win comes from skipping over big blocks of zeros 32/64
bits at a time, but it was neat to use the ffsl function in there to take
advantage of hardware instructions to find the next set bit (i.e. bsf on
x86).

Cleaned up the proliforation of "unsigned long" throughout the code and
replaced with the leafblock_t typedef.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>