win: Try to work around unstable timers on buggy multicore setups
Apparently there are still some problems with the performance counters on
some Windows systems due to BIOS or driver bugs.

 - http://www.quaddicted.com/forum/viewtopic.php?pid=405#p405
 - http://forums.inside3d.com/viewtopic.php?p=48051#p48051
 - http://forums.inside3d.com/viewtopic.php?p=36239

Found the MS recommendations for this, which says to set CPU affinity...
not so keen on that for now, so let's implement the workaround written up
by MH (thanks!)

Thanks Spirit (http://www.quaddicted.com/) for pointing out this problem.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>