render: use a separate pass over the BSP tree for frustum culling
Pre-calculate frustum clip flags for each node/surface in the PVS by using
a separate walk of the tree before drawing. Doesn't seem to be any
measureable performance benefit, but it was nice to separate this step
from the drawing.

The BSP walk doesn't need to be ordered, so could be done iteratively with
no node stack.

I'm hoping I can apply this to submodels too without too much performance
impact for small models and with some benefit for large, complex models...

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>