vid: convert vid width, height, rowbytes to signed quantities
Mostly important for the rowbytes, which are negative for a bottom-up
windows DIB. Big thanks to szo for posting about this gotcha to the forum.

Also had to change a few places where a bit shift was used to divide
rowbytes by 2. Behaviour is undefined for signed numbers with negative
values.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>