WIIU: fix the build
== DETAILS
The build is broken for a couple of reasons:

1. The wiiu target sets flags for GameCube/Wii that should not be set
2. Even though `retro_sleep()` isn't actually called, its include is
   still present and that causes the build to fail on wiiu

The `retro_sleep` compile failure is kinda its own issue, but in this
case we can sidestep it. The proper fix for this would be to update
this core so it doesn't keep its own copy of libretro-common, but that's
more work than I have time to complete.

Change summary:

1. remove the GC/Wii flags so we don't try to include libogc headers
2. remove `retro_sleep()` and go a little further and remove the
   `Sys_Sleep()` function that isn't actually called anywhere.

I have confirmed that it builds. I have not confirmed that it works.