Fix HUD flash on damage, ammo pickup, etc. (#138)
When the screen flashes (e.g. the player takes damage), the HUD sometimes gets stuck with the flash hue until something else forces it to be updated. Also, the flash hue is supposed to decay, and that almost never happens to the HUD under lr-tyrquake.

The HUD needs to be marked dirty when the screen palette is shifted, to force it to be redrawn even if nothing else has changed in it.